requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
google-generativeai
google-auth
google-api-python-client
supabase
python-dotenv
pypdf
python-docx
openpyxl
pydantic
httpx
requests
numpy